Interface

MenuAdventures.VERB_FOR

MenuAdventures.AbstractDoor

MenuAdventures.AbstractRoom

MenuAdventures.Action

MenuAdventures.Attach

MenuAdventures.Close

MenuAdventures.Direction

MenuAdventures.Domain

MenuAdventures.Door

MenuAdventures.Dress

MenuAdventures.Drop

MenuAdventures.Eat

MenuAdventures.ExitDirections

MenuAdventures.Give

MenuAdventures.Go

MenuAdventures.GoInto

MenuAdventures.GoOnto

MenuAdventures.GrammaticalPerson

MenuAdventures.Inventory

MenuAdventures.Leave

MenuAdventures.ListInventory

MenuAdventures.Location

MenuAdventures.Lock

MenuAdventures.LookAt

MenuAdventures.MoveSiblings

MenuAdventures.Noun

MenuAdventures.Open

MenuAdventures.Push

MenuAdventures.PutInto

MenuAdventures.PutOnto

MenuAdventures.Quit

MenuAdventures.Reachable

MenuAdventures.Relationship

MenuAdventures.Take

MenuAdventures.TurnOff

MenuAdventures.TurnOn

MenuAdventures.Universe

MenuAdventures.Unlock

MenuAdventures.Verb

MenuAdventures.Visible

MenuAdventures.Wear

MenuAdventures.argument_domains

MenuAdventures.blocking

MenuAdventures.ever_possible

MenuAdventures.get_description

MenuAdventures.is_providing_light

MenuAdventures.is_transparent

MenuAdventures.is_vehicle

MenuAdventures.opposite

MenuAdventures.possible_now

MenuAdventures.possible_now

MenuAdventures.print_sentence

MenuAdventures.subject_to_verb

MenuAdventures.turn!

MenuAdventures.VERB_FOR — Constant

const VERB_FOR = Dict{Relationship, Verb}

Get the verb form of a Relationship, or roughly, remove the ing.

MenuAdventures.AbstractDoor — Type

An abstract door

MenuAdventures.AbstractRoom — Type

abstract type AbstractRoom <: Location end

Must have a mutable visited field.

MenuAdventures.Action — Type

abstract type Action end

A action is an action the player can take.

It should be fairly easy to create new verbs: you will need to define ever_possible for abstract possibilities, possible_now for concrete possibilities, argument_domains to specify the domain of the arguments, and print_sentence for printing the sentence.

Most importantly, define:

function (::MyNewAction)(universe, arguments...) -> Bool

Which will conduct the action based on user choices. Return true to end the game.

MenuAdventures.Attach — Type

struct Attach <: Action end

Attach something to something else.

MenuAdventures.Close — Type

struct Close <: Action end

Close something.

MenuAdventures.Direction — Type

@enum Direction north south west east north_west north_east south_west south_east up down inside outside

Directions show the relationships between Locations.

You can use opposite to find the opposite of a direction.

MenuAdventures.Domain — Type

abstract type Domain end

A domain refers to a search space for a specific argument to a action.

For example, you are only able to look at things in the Visible domain. Domains serve both as a way of distinguishing different arguments to a action, and also, categorizing the environment around the player. Users could theoretically add a new domain.

MenuAdventures.Door — Type

A door

MenuAdventures.Dress — Type

struct Dress <: Action end

Dress someone in something.

MenuAdventures.Drop — Type

struct Drop <: Action end

Drop something.

MenuAdventures.Eat — Type

struct Eat <: Action end

Eat something.

MenuAdventures.ExitDirections — Type

struct ExitDirections <: Domain end

Directions that a player, or the vehicle a player is in, might exit in.

MenuAdventures.Give — Type

struct Give <: Action end

Give something to someone.

MenuAdventures.Go — Type

struct Go <: Action end

Go some way.

MenuAdventures.GoInto — Type

struct GoInto <: Action end

Go into something.

MenuAdventures.GoOnto — Type

struct GoOnto <: Action end

Go onto something.

MenuAdventures.GrammaticalPerson — Type

@enum GrammaticalPerson first_person second_person third_person

Grammatical person

MenuAdventures.Inventory — Type

struct Inventory <: Domain end

Things the player is carrying.

MenuAdventures.Leave — Type

struct Leave <: Action end

Leave something.

MenuAdventures.ListInventory — Type

struct ListInventory <: Action end

List your inventory.

MenuAdventures.Location — Type

abstract type Location <: Noun end

A location (room or door)

MenuAdventures.Lock — Type

struct Lock <: Action end

Lock something with something.

MenuAdventures.LookAt — Type

struct LookAt <: Action end

Look at something.

MenuAdventures.MoveSiblings — Type

struct MoveSiblings <: Domain end

Thing that are in/on the same place the player could more from.

MenuAdventures.Noun — Type

abstract type Noun end

Nouns must have the following fields:

name::String plural::Bool grammaticalperson::GrammaticalPerson indefinitearticle::String

They are characterized by the following traits and methods:

ever_possible

get_description,

is_providing_light,

is_transparent,

is_vehicle.

Set indefinite_article to "" for proper nouns.

MenuAdventures.Open — Type

struct Open <: Action end

Open something.

MenuAdventures.Push — Type

struct Push <: Action end

Push something some way.

MenuAdventures.PutInto — Type

struct PutInto <: Action end

Put something into something.

MenuAdventures.PutOnto — Type

struct PutOnto <: Action end

PutOnto something onto something

MenuAdventures.Quit — Type

struct Quit <: Action end

Quit

MenuAdventures.Reachable — Type

struct Reachable <: Domain end

Anything the player possible_now reach.

Players possible_now't reach through closed containers by default.

MenuAdventures.Relationship — Type

@enum Relationship carrying containing incorporating supporting wearing

Relationships show the relationshp of a thing to its parent_thing.

A is containing B means B is in A A is incorporating B means B is part of A A is supporting B means B is on top of A

MenuAdventures.Take — Type

struct Take <: Action end

Take something

MenuAdventures.TurnOff — Type

struct TurnOff <: Action end

Turn something off

MenuAdventures.TurnOn — Type

struct TurnOn <: Action end

Turn something on

MenuAdventures.Universe — Method

function Universe(
    player;
    interface = terminal,
    introduction = "",
    relationships_graph = MetaGraph(DiGraph(), Label = Noun, EdgeMeta = Relationship),
    directions_graph = MetaGraph(DiGraph(), Label = Location, EdgeMeta = Direction),
    choices_log::Vector{Int}
)

The universe contains a player, a text interface, an introduction, and the relationships between nouns and locations.

The universe is organized as interlinking web of locations connected by Directions. For any origin and destination, there should be no more than one connection to a particular direction. Each location is the root of a Relationship tree. Every noun should have one and only one parent, (except for locations), which are at the root of trees and have no parent.

You can add a new thing to the universe, or change its location, by specifying its relation to another thing:

universe[parent_thing, thing, silent = false] = relationship

Set silent = true to suppress the "Ok" message.

You possible_now add a connection between locations too, optionally interspersed by a door:

universe[parent_thing, destination, one_way = false] = direction
universe[parent_thing, destination, one_way = false] = door, direction

By default, this will create a way back in the opposite direction. To suppress this, set one_way = true

MenuAdventures.Unlock — Type

struct Unlock <: Action end

Unlock something

MenuAdventures.Verb — Method

function Verb(base; third_person_singular_present = string(base, "s"))

Create an English verb.

Use subject_to_verb to get the form of a verb to agree with a subject.

MenuAdventures.Visible — Type

struct Visible <: Domain end

Anything the player possible_now see.

Players possible_now't see into closed, opaque containers.

MenuAdventures.Wear — Type

struct Wear <: Action end

Wear something.

MenuAdventures.argument_domains — Method

function argument_domains(action::Action)

A tuple of the Domains for each argument of an action.

MenuAdventures.blocking — Method

blocking(domain, parent_thing, relationship, thing)

parent_thing is blocked from accessing thing via the relationship.

By default, Reachable parent_things block things they are containing if they are closed. By default, Visible parent_things block things they are containing if they are closed and not is_transparent.

MenuAdventures.ever_possible — Method

ever_possible(action::Action, domain::Domain, noun::Noun)

Whether is is abstractly possible to apply a action to a noun from a particular domain.

For whether it is concretely possible for the player in at a certain moment, see possible_now. Most possibilities default to false, with some exceptions:

ever_possible(::PutInto, ::Inventory, _) = true
ever_possible(::Drop, ::Inventory, _) = true
ever_possible(::PutOnto, ::Inventory, _) = true
ever_possible(action::TurnOff, domain::Reachable, noun) = 
    ever_possible(TurnOn(), domain, noun)
ever_possible(action::Close, domain::Reachable, noun) = 
    ever_possible(Open(), domain, noun)
ever_possible(action::Lock, domain::Reachable, noun) = 
    ever_possible(Unlock(), domain, noun)
ever_possible(::Lock, domain::Inventory, noun) = 
    ever_possible(Unlock(), domain, noun)

Certain possibilities come with required fields:

ever_possible(::TurnOn, ::Reachable, noun) requires that noun has a mutable on::Bool field. ever_possible(::Open, ::Reachable, noun requires that noun has a mutable closed::Bool field. ever_possible(::Unlock, ::Reachable, noun) requires that noun has a key::Noun field and a mutable locked::Bool field. ever_possible(::Take, ::Reachable, noun) requires that noun has a mutable handled::Bool field.

MenuAdventures.get_description — Method

get_description(universe, thing::Noun) = thing.description

Get the description of a thing.

Unless you overload get_description, nouns are required to have a description field.

MenuAdventures.is_providing_light — Method

is_providing_light(::Noun) = false

Whether something provides its own light. Naturally lit locations and light sources both are providing light.

MenuAdventures.is_transparent — Method

is_transparent(thing::Noun) = false

Whether you can see through thing into its contents.

MenuAdventures.is_vehicle — Method

is_vehicle(::Noun) = false

Whether something is a vehicle.

MenuAdventures.opposite — Method

function opposite(direction::Direction)

The opposite of a direction.

MenuAdventures.possible_now — Method

possible_now(universe, sentence, domain, thing)

Whether it is currently ever_possible to apply sentence.action to a thing in a domain.

See ever_possible for a more abstract possibility. sentence will contain already chosen arguments should you wish to access them.

MenuAdventures.possible_now — Method

possible_now(universe, action)

Whether it is possible to conduct an action. Defaults to true; you can set to false for some actions without arguments.

MenuAdventures.print_sentence — Method

function print_sentence(io, action::Action, argument_texts...)

Print a sentence to io. This allows for adding connectives like with.

MenuAdventures.subject_to_verb — Method

subject_to_verb(subject, verb)

Find the Verb form to agree with a subject.

MenuAdventures.turn! — Method

turn!(universe; introduce = false, should_look_around = false)

Start a turn in the Universe, and keep going until the user wins or quits.