Internals

MenuAdventures.argument_domains

MenuAdventures.blocking

MenuAdventures.ever_possible

MenuAdventures.find_in_domain

MenuAdventures.indefinite

MenuAdventures.interrogative

MenuAdventures.is_shining

MenuAdventures.is_transparent

MenuAdventures.is_vehicle

MenuAdventures.look_around

MenuAdventures.mention_status

MenuAdventures.opposite

MenuAdventures.possible_now

MenuAdventures.possible_now

MenuAdventures.print_sentence

MenuAdventures.argument_domains — Function

argument_domains(action::Action)

Return a tuple of the Domains for each argument of an Action.

MenuAdventures.blocking — Function

MenuAdventures.blocking(domain, parent_thing, relationship, thing)

parent_thing is blocked from accessing thing via the relationship.

By default, Reachable parent_things block things they are containing if they are closed. By default, Visible parent_things block things they are containing if they are closed and not MenuAdventures.is_transparent.

MenuAdventures.ever_possible — Method

MenuAdventures.ever_possible(action::Action, domain::Domain, noun::Noun)

Whether it is abstractly possible to apply an Action to a Noun from a particular Domain.

For whether it is concretely possible for the player in at a certain moment, see possible_now. Most possibilities default to false, with some exceptions, documented in specific actions.

Certain possibilities come with required fields:

ever_possible(::Open, ::Reachable, noun requires that noun has a mutable closed::Bool field.

ever_possible(::Unlock, ::Reachable, noun) requires that noun has a key::Noun field and a mutable locked::Bool field.

MenuAdventures.find_in_domain — Function

MenuAdventures.find_in_domain(universe, sentence, domain; lit = true)

Return a vector of Answers for all things around the player in a certain Domain.

MenuAdventures.indefinite — Method

MenuAdventures.indefinite(domain) = "something"

Placeholder for an object from the Domain.

For example, the indefinite for ExitDirections is "some way".

MenuAdventures.interrogative — Method

MenuAdventures.interrogative(domain) = "something"

Ask for an object from the Domain.

For example, the interrogative for ExitDirections is "which way".

MenuAdventures.is_shining — Method

is_shining(::Noun) = false

Whether something provides its own light.

MenuAdventures.is_transparent — Method

MenuAdventures.is_transparent(thing::Noun) = false

Whether you can see through thing into its contents.

MenuAdventures.is_vehicle — Method

MenuAdventures.is_vehicle(::Noun) = false

Whether something is a vehicle.

MenuAdventures.look_around — Method

MenuAdventures.look_around(universe, domain, blocking_thing, blocked_relationship)

Look around a lit location.

You can overload look_around for a Noun subtype. Use Core.invoke to avoid replicating the look_around machinery.

MenuAdventures.mention_status — Method

mention_status(io, action, thing)

Print the status of thing corresponding to action into io.

For example, for the OpenOrClose action, will display whether thing is open or closed. Called for all action subtypes when mentioning a thing. Defaults to doing nothing.

MenuAdventures.opposite — Function

MenuAdventures.opposite(direction::Direction)

The opposite of a Direction.

MenuAdventures.possible_now — Method

MenuAdventures.possible_now(universe, sentence, domain, thing)

Whether it is currently possible to apply sentence.action to a thing in a domain.

See ever_possible for a more abstract possibilities. sentence will contain already chosen arguments, should you wish to access them.

MenuAdventures.possible_now — Method

possible_now(universe, action)

Whether it is possible to conduct an action. Defaults to true; you can set to false for some actions without arguments.

MenuAdventures.print_sentence — Function

MenuAdventures.print_sentence(io, action::Action, argument_answers::Answer...)

Print a sentence to io.

This allows for inserting connectives like with. Arguments will be passed as Answers.