Interface

MenuAdventures.AbstractDoor

MenuAdventures.AbstractRoom

MenuAdventures.Action

MenuAdventures.Answer

MenuAdventures.Attach

MenuAdventures.Close

MenuAdventures.Dialog

MenuAdventures.Direction

MenuAdventures.Domain

MenuAdventures.Dress

MenuAdventures.Drop

MenuAdventures.Eat

MenuAdventures.ExitDirections

MenuAdventures.Give

MenuAdventures.Go

MenuAdventures.GoInto

MenuAdventures.GoOnto

MenuAdventures.GrammaticalPerson

MenuAdventures.Immediate

MenuAdventures.Inventory

MenuAdventures.Leave

MenuAdventures.ListInventory

MenuAdventures.Location

MenuAdventures.Lock

MenuAdventures.LookAround

MenuAdventures.LookAt

MenuAdventures.Noun

MenuAdventures.Open

MenuAdventures.Outfit

MenuAdventures.Push

MenuAdventures.PutInto

MenuAdventures.PutOnto

MenuAdventures.Quit

MenuAdventures.Reachable

MenuAdventures.Relationship

MenuAdventures.Say

MenuAdventures.Sentence

MenuAdventures.Take

MenuAdventures.TakeOff

MenuAdventures.TurnOff

MenuAdventures.TurnOn

MenuAdventures.Universe

MenuAdventures.Unlock

MenuAdventures.Verb

MenuAdventures.Visible

MenuAdventures.Wear

MenuAdventures.argument_domains

MenuAdventures.blocking

MenuAdventures.blocking_thing_and_relationship

MenuAdventures.ever_possible

MenuAdventures.get_children_relationships

MenuAdventures.get_description

MenuAdventures.get_exit_directions

MenuAdventures.get_final_destination

MenuAdventures.get_first_destination

MenuAdventures.get_mover

MenuAdventures.get_parent

MenuAdventures.get_parent_relationship

MenuAdventures.is_providing_light

MenuAdventures.is_transparent

MenuAdventures.is_vehicle

MenuAdventures.opposite

MenuAdventures.possible_now

MenuAdventures.possible_now

MenuAdventures.print_sentence

MenuAdventures.string_in_color

MenuAdventures.subject_to_verb

MenuAdventures.turn!

MenuAdventures.AbstractDoor — Type

An abstract door

MenuAdventures.AbstractRoom — Type

abstract type AbstractRoom <: Location end

Must have a mutable visited::Bool field.

MenuAdventures.Action — Type

abstract type Action end

A action is an action the player can take.

It should be fairly easy to create new verbs: you will need to define

ever_possible for abstract possibilities,

possible_now for concrete possibilities,

argument_domains to specify the domain of the arguments

print_sentence for printing the sentence. Note that the order arguments are printed need not match the order they are listed.

Most importantly, define:

function (::MyNewAction)(universe, arguments...) -> Bool

Which will conduct the action based on user choices. Return true to end the game.

MenuAdventures.Answer — Type

struct Answer

An answer has two fields: text, which will be how the option is displayed in a menu, and object. object might be a noun, direction, trigger, or even a question.

MenuAdventures.Attach — Type

struct Attach <: Action end

Attach something from your Inventory to something else Reachable.

MenuAdventures.Close — Type

struct Close <: Action end

Close something Reachable.

MenuAdventures.Dialog — Type

struct Dialog <: Domain end

Things a player might say.

A Domain.

MenuAdventures.Direction — Type

@enum Direction north south west east north_west north_east south_west south_east up down inside outside

Directions show the relationships between Locations.

You can use opposite to find the opposite of a direction.

MenuAdventures.Domain — Type

abstract type Domain end

A domain refers to a search space for a specific argument to an Action.

For example, you are only able to look at things in the Visible domain. Domains serve both as a way of distinguishing different arguments to an action, and also, categorizing the environment around the player. Users could theoretically add a new domain.

MenuAdventures.Dress — Type

struct Dress <: Action end

Dress someone Reachable in something from your Inventory.

MenuAdventures.Drop — Type

struct Drop <: Action end

Drop something from your Inventory.

MenuAdventures.Eat — Type

struct Eat <: Action end

Eat something Reachable.

MenuAdventures.ExitDirections — Type

struct ExitDirections <: Domain end

Directions that a player, or the vehicle a player is in, might exit in.

A Domain.

MenuAdventures.Give — Type

struct Give <: Action end

Give something from your Inventory to someone Reachable.

MenuAdventures.Go — Type

struct Go <: Action end

Go in one of ExitDirections.

MenuAdventures.GoInto — Type

struct GoInto <: Action end

Go into something Immediate.

MenuAdventures.GoOnto — Type

struct GoOnto <: Action end

Go onto something Immediate.

MenuAdventures.GrammaticalPerson — Type

@enum GrammaticalPerson first_person second_person third_person

Grammatical person

MenuAdventures.Immediate — Type

struct Immediate <: Domain end

Thing that are in/on the same place the player could more from.

A Domain.

MenuAdventures.Inventory — Type

struct Inventory <: Domain end

Things the player is carrying.

A Domain.

MenuAdventures.Leave — Type

struct Leave <: Action end

Leave whatever the player is in/on.

MenuAdventures.ListInventory — Type

struct ListInventory <: Action end

List the player's inventory.

MenuAdventures.Location — Type

abstract type Location <: Noun end

A location (room or door)

MenuAdventures.Lock — Type

struct Lock <: Action end

Lock something Reachable with something from your Inventory.

MenuAdventures.LookAround — Type

struct LookAround <: Action end

Look around.

MenuAdventures.LookAt — Type

struct LookAt <: Action end

Look at something Visible.

MenuAdventures.Noun — Type

abstract type Noun end

Nouns must have the following fields:

name::String

plural::Bool

grammatical_person::GrammaticalPerson (see GrammaticalPerson)

indefinite_article::String

They are characterized by the following traits and methods:

ever_possible

get_description,

is_providing_light,

is_transparent,

is_vehicle.

Set indefinite_article to "" for proper nouns. The following IOContext components will be respected when showing nouns:

:capitalize::Bool => false

:known::Bool => true, set to false to include the indefinite article if it exists.

:is_subject => true, whether the noun is the subject of the sentence. If this is set to false, you must also include

:subject::Noun, the subject of the sentence.

There are a handful of unexported Noun sub-types included mostly for testing purposes:

MenuAdventures.Box

MenuAdventures.Car

MenuAdventures.Clothes

MenuAdventures.Door

MenuAdventures.Key

MenuAdventures.Lamp

MenuAdventures.Person

MenuAdventures.StickyThing

MenuAdventures.Table

Feel free to use them if you want, but for the most part, you will likely want to make your own custom Noun subtypes for almost everything in your game.

MenuAdventures.Open — Type

struct Open <: Action end

Open something Reachable.

MenuAdventures.Outfit — Type

struct Outfit <: Domain end

Things the player is wearing.

A Domain.

MenuAdventures.Push — Type

struct Push <: Action end

Push something Immediate in one of ExitDirections.

MenuAdventures.PutInto — Type

struct PutInto <: Action end

Put something from your Inventory into something Reachable.

MenuAdventures.PutOnto — Type

struct PutOnto <: Action end

Put something from your Inventory onto something Reachable.

MenuAdventures.Quit — Type

struct Quit <: Action end

Quit

MenuAdventures.Reachable — Type

struct Reachable <: Domain end

Anything the player can reach.

A Domain. Players can't reach through closed containers by default.

MenuAdventures.Relationship — Type

@enum Relationship carrying containing incorporating supporting wearing

Relationships show the relationshp of a thing to its parent_thing.

A is carrying B means B is carried by A

A is containing B means B is in A

A is incorporating B means B is part of A

A is supporting B means B is on top of A

A is wearing B means B is worn by A

MenuAdventures.Say — Type

struct Say <: Action end

Say, to someone Visible, some Dialog.

MenuAdventures.Sentence — Type

struct Sentence{Action}

A sentence has two fields: action, the Action to tbe taken, and arguments, the arguments to the action.

The subject is implicitly universe.player.

MenuAdventures.Take — Type

struct Take <: Action end

Take something Reachable.

MenuAdventures.TakeOff — Type

struct TakeOff <: Action end

Take off part of your Outfit.

MenuAdventures.TurnOff — Type

struct TurnOff <: Action end

Turn off something Reachable.

MenuAdventures.TurnOn — Type

struct TurnOn <: Action end

Turn on something Reachable.

MenuAdventures.Universe — Method

function Universe(
    player;
    interface = terminal,
    introduction = "",
    relationships_graph = MetaGraph(DiGraph(), Label = Noun, EdgeMeta = Relationship),
    directions_graph = MetaGraph(DiGraph(), Label = Location, EdgeMeta = Direction),
    choices_log::Vector{Int}
)

The universe contains a player, a text interface, an introduction, and the relationships between Nouns and Locations.

The player will typically have second_person as their GrammaticalPerson. The universe is organized as interlinking web of locations connected by Directions. Each location is the root of a Relationship tree. Locations are at the root of trees and have no parent. Every other noun should have one and only one parent in the relationship tree.

You can add a new thing to the universe, or change the location of something, by specifying its relation to another thing:

universe[parent_thing, thing] = relationship

You can add a connection between locations too, optionally interspersed by a door:

universe[parent_thing, destination, one_way = false] = direction
universe[parent_thing, destination, one_way = false] = door, direction

By default, this will create a way back in the opposite direction. To suppress this, set one_way = true

MenuAdventures.Unlock — Type

struct Unlock <: Action end

Unlock something Reachable with something from your Inventory.

MenuAdventures.Verb — Method

function Verb(base; third_person_singular_present = string(base, "s"))

Create an English verb.

Use subject_to_verb to get the form of a verb to agree with a subject. Unexported verbs include MenuAdventures.DO, MenuAdventures.BE, and MenuAdventures.REPLY.

MenuAdventures.Visible — Type

struct Visible <: Domain end

Anything the player can see.

A Domain. By default, players can't see into closed, opaque containers.

MenuAdventures.Wear — Type

struct Wear <: Action end

Wear something from your Inventory.

MenuAdventures.argument_domains — Method

function argument_domains(action::Action)

Return a tuple of the Domains for each argument of an Action.

MenuAdventures.blocking — Method

blocking(domain, parent_thing, relationship, thing)

parent_thing is blocked from accessing thing via the relationship.

By default, Reachable parent_things block things they are containing if they are closed. By default, Visible parent_things block things they are containing if they are closed and not is_transparent.

MenuAdventures.blocking_thing_and_relationship — Method

function blocking_thing_and_relationship(universe, domain)

Starting from the player, search upwards in the Relationships tree until a blocking relationship is reached, or until the root of the tree is reached. Return the blocking thing and the blocked relationship.

MenuAdventures.ever_possible — Method

ever_possible(action::Action, domain::Domain, noun::Noun)

Whether it is abstractly possible to apply an Action to a Noun from a particular Domain.

For whether it is concretely possible for the player in at a certain moment, see possible_now. Most possibilities default to false, with some exceptions:

ever_possible(::Drop, ::Inventory, _) = true
ever_possible(::Give, ::Inventory, _) = true
ever_possible(::Attach, ::Reachable, _) = true
ever_possible(::PutOnto, ::Inventory, _) = true
ever_possible(::PutInto, ::Inventory, _) = true
ever_possible(::TakeOff, ::Outfit, _) = true

ever_possible(::Dress, domain::Inventory, noun) = ever_possible(Wear(), domain, noun)
ever_possible(::TurnOff, domain::Reachable, noun) = ever_possible(TurnOn(), domain, noun)
ever_possible(::Close, domain::Reachable, noun) = ever_possible(Open(), domain, noun)
ever_possible(::Lock, domain::Reachable, noun) = ever_possible(Unlock(), domain, noun)
ever_possible(::Lock, domain::Inventory, noun) = ever_possible(Unlock(), domain, noun)

Certain possibilities come with required fields:

ever_possible(::TurnOn, ::Reachable, noun) requires that noun has a mutable on::Bool field.

ever_possible(::Open, ::Reachable, noun requires that noun has a mutable closed::Bool field.

ever_possible(::Unlock, ::Reachable, noun) requires that noun has a key::Noun field and a mutable locked::Bool field.

ever_possible(::Take, ::Reachable, noun) requires that noun has a mutable handled::Bool field.

ever_possible(::Say, ::Visible, noun) requires that noun have a dialog::OrderedDict{String, Answer} field.

ever_possible(::GoInto, ::Immediate, noun) and ever_possible(::GoOnto, ::Immediate, noun) require than noun has a mutable visited::Bool field.

For dialog, the keys will be things you can say, and answers will be replies, with the answer objects being function triggers that will be called with no arguments. Dialog options will be removed after usage.

MenuAdventures.get_children_relationships — Method

function get_children_relationships(universe, parent_thing)

Get the children of parent_thing in the universe, and the Relationships of parent_thing to them.

MenuAdventures.get_description — Method

get_description(universe, thing::Noun) = thing.description

Get the description of a thing.

Unless you overload get_description, nouns are required to have a description field.

MenuAdventures.get_exit_directions — Method

function get_exit_directions(universe, location)

Get the exits from a location, and the Directions to them.

MenuAdventures.get_final_destination — Method

function get_final_destination(universe, direction)

Get the final destination if universe.player moves in direction. Players always continue through doors to a new final destination.

MenuAdventures.get_first_destination — Method

function get_first_destination(universe, direction)

Get the first destination if universe.player moves in direction (a location or a door).

MenuAdventures.get_mover — Method

function get_mover(universe)

Return the thing that can move: either the player, or the vehicle the player is in, if it is_vehicle.

MenuAdventures.get_parent — Method

function get_parent(universe, thing)

Get the parent of thing in the Relationship tree.

MenuAdventures.get_parent_relationship — Method

function get_parent_relationship(universe, thing)

Get the parent of a thing in the universe, and the parent's Relationship to it.

MenuAdventures.is_providing_light — Method

is_providing_light(::Noun) = false

Whether something provides its own light. Naturally lit locations and light sources both are providing light.

MenuAdventures.is_transparent — Method

is_transparent(thing::Noun) = false

Whether you can see through thing into its contents.

MenuAdventures.is_vehicle — Method

is_vehicle(::Noun) = false

Whether something is a vehicle.

MenuAdventures.opposite — Method

function opposite(direction::Direction)

The opposite of a direction.

MenuAdventures.possible_now — Method

possible_now(universe, sentence, domain, thing)

Whether it is currently possible to apply sentence.action to a thing in a domain.

See ever_possible for a more abstract possibility. sentence will contain already chosen arguments, should you wish to access them.

MenuAdventures.possible_now — Method

possible_now(universe, action)

Whether it is possible to conduct an action. Defaults to true; you can set to false for some actions without arguments.

MenuAdventures.print_sentence — Method

function print_sentence(io, action::Action, argument_texts...)

Print a sentence to io. This allows for adding connectives like with.

MenuAdventures.string_in_color — Method

function string_in_color(color::Symbol, arguments...)

Use ASCII escape codes add a color to the arguments collected as a string.

MenuAdventures.subject_to_verb — Method

subject_to_verb(subject, verb)

Find the Verb form to agree with a subject.

MenuAdventures.turn! — Method

turn!(universe; introduce = false)

Start a turn in the Universe, and keep going until an Action returns true.